Massive equipmentParser refactor PT. 1
