+++ scripts/utils/equipment/index.js
@@ -1,0 +1,28 @@
+/** Equipment module for Hero Mancer character creation */
+
+/** Main equipment data fetching and processing service */
+export { EquipmentDataService } from './equipmentDataService.js';
+
+/** Core equipment parser class that coordinates rendering and data services */
+export { EquipmentParser } from './equipmentParser.js';
+
+/** Manages rendering of equipment selection UI components */
+export { EquipmentRenderer } from './equipmentRenderer.js';
+
+/** Renderer for AND-type equipment blocks (combined items) */
+export { AndItemRenderer } from './renderers/andItemRenderer.js';
+
+/** Base renderer class with shared equipment rendering functionality */
+export { BaseItemRenderer } from './renderers/baseItemRenderer.js';
+
+/** Renderer for arcane/divine focus equipment items */
+export { FocusItemRenderer } from './renderers/focusItemRenderer.js';
+
+/** Renderer for linked individual equipment items */
+export { LinkedItemRenderer } from './renderers/linkedItemRenderer.js';
+
+/** Renderer for OR-type equipment choice blocks */
+export { OrItemRenderer } from './renderers/orItemRenderer.js';
+
+/** Renderer for tool equipment items */
+export { ToolItemRenderer } from './renderers/toolItemRenderer.js';
